Refactor cache code to own module
